Added support for named PCI slots.
A new string property, "slot", is introduced for PCI devices;
a new syntax is introduced for the "addr" property of PCI devices.

A function of a multifunction PCI device can now specify
just the function number, following a dot, in the "addr" property.
For example, ".1".

Each function of the same PCI device will then specify
the "slot" property, naming the device's slot in the bus.